AGS-speakNotes

+++ controller/fileController.js
@@ -2,7 +2,7 @@ import path from "path";
 import { spawn } from "child_process";
 import fs from "fs";
 import { convertToWav } from "../Middlewares/convertToWav.js";
-
+import { createUser, selectAudio, selectUser } from "../models/sql.js";
 
 export const uploadController = async (req, res) => {
   const audio = req.file;
@@ -10,6 +10,7 @@ export const uploadController = async (req, res) => {
   const wavFile = `${audio.originalname}.wav`;
   const wavPath = path.resolve("upload", wavFile);
   const ffmpegPath = "C:/ffmpeg/ffmpeg-7.1.1-essentials_build/bin/ffmpeg.exe";
+  const {Id, displayName} = req.body;
 
   try {
     
@@ -25,9 +26,19 @@ export const uploadController = async (req, res) => {
       console.error("Python error:", err.toString());
     });
 
-    py.on("close", () => {
+    py.on("close", async () => {
+      const insert_audio = {
+        user_Id: Id,
+        wav_file: wavFile,
+        result_text: result.trim(),
+        display_name: displayName
+      }
+      const insert_result = await createUser("tblaudio", insert_audio);
+      const select_results = await selectAudio(insert_result.insertId);
+      const select_user_result = await selectUser(displayName);
+      if(!select_results  || !select_user_result) return res.render("Home", {msg: "something went wrong!!!!", profile: select_user_result});
       fs.unlinkSync(inputPath);
-      return res.render("testing", { result: result.trim() });
+      return res.render("Home", { profile: select_user_result, transcriptions: select_results});
     });
   } catch (error) {
     console.log(error);
